CORE-6350,CORE-6332 Begin details view refactor
