+++ modules/themes/src/main/java/org/iplantc/de/theme/base/client/diskResource/details/DetailsViewDefaultAppearance.java
@@ -1,0 +1,10 @@
+package org.iplantc.de.theme.base.client.diskResource.details;
+
+import org.iplantc.de.diskResource.DetailsView;
+
+/**
+ * @author jstroot
+ */
+public class DetailsViewDefaultAppearance implements DetailsView.Appearance {
+
+}
