+++ modules/themes/src/main/java/org/iplantc/de/theme/base/client/diskResource/grid/GridViewDisplayStrings.java
@@ -36,6 +36,9 @@ public interface GridViewDisplayStrings extends Messages{
     @Key("removeFavoriteError")
     String removeFavoriteError();
 
+    @Key("retrieveStatError")
+    String retrieveStatError();
+
     @Key("share")
     String share();
