+++ modules/themes/src/main/java/org/iplantc/de/theme/base/client/diskResource/grid/GridViewDisplayStrings.properties
@@ -10,4 +10,4 @@ metadataFormInvalid = Please fix all errors before saving.
 metadataHelp = <p>An attribute is a changeable property or characteristic of the file or folder you have selected that can be set to a value.</p>  <p>For example, "filetype" is an attribute of a file and "text" would be its value.</p>  <p>By entering this information, you are describing the contents and context of the data. This can be very useful when you are trying to remember specific information about that data in the future, or sharing a file so collaborators may know the context of the data, and many other instances.</p>
 markFavoriteError = Unable to mark selected item as favorite.
 removeFavoriteError = Unable to remove selected item from favorites.
-
+retrieveStatError =  Unable to retrieve details.
