+++ modules/ui-disk-resource-module/src/main/java/org/iplantc/de/diskResource/client/GridView.java
@@ -4,6 +4,7 @@ import org.iplantc.de.client.models.HasId;
 import org.iplantc.de.client.models.diskResources.DiskResource;
 import org.iplantc.de.client.models.diskResources.Folder;
 import org.iplantc.de.diskResource.client.events.*;
+import org.iplantc.de.diskResource.client.events.selection.EditInfoTypeSelected;
 import org.iplantc.de.diskResource.client.events.selection.ManageCommentsSelected;
 import org.iplantc.de.diskResource.client.events.selection.ManageMetadataSelected;
 import org.iplantc.de.diskResource.client.events.selection.ManageSharingSelected;
@@ -17,6 +18,7 @@ import com.google.gwt.event.shared.EventHandler;
 import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.user.client.ui.IsWidget;
 
+import com.sencha.gxt.data.shared.event.StoreUpdateEvent;
 import com.sencha.gxt.data.shared.loader.BeforeLoadEvent;
 import com.sencha.gxt.data.shared.loader.PagingLoadResult;
 import com.sencha.gxt.data.shared.loader.PagingLoader;
@@ -90,7 +92,9 @@ public interface GridView extends IsWidget,
                                 RequestDiskResourceFavoriteEvent.RequestDiskResourceFavoriteEventHandler,
                                 ManageCommentsSelected.ManageCommentsSelectedEventHandler,
                                 FolderSelectionEvent.FolderSelectionEventHandler,
-                                SubmitDiskResourceQueryEvent.SubmitDiskResourceQueryEventHandler {
+                                SubmitDiskResourceQueryEvent.SubmitDiskResourceQueryEventHandler,
+                                StoreUpdateEvent.HasStoreUpdateHandlers<DiskResource>,
+                                EditInfoTypeSelected.EditInfoTypeSelectedEventHandler {
 
         interface Appearance {
 
@@ -119,6 +123,8 @@ public interface GridView extends IsWidget,
             String metadataHelp();
 
             String removeFavoriteError();
+
+            String retrieveStatFailed();
 
             String shareLinkDialogHeight();
 
@@ -149,13 +155,10 @@ public interface GridView extends IsWidget,
 
         void setParentPresenter(DiskResourceView.Presenter parentPresenter);
 
-        void setSelectedDiskResources(List<? extends HasId> diskResourcesToSelect);
-
         void setSelectedDiskResourcesById(List<? extends HasId> diskResourcesToSelect);
 
         void unRegisterHandler(EventHandler handler);
 
-        DiskResource updateDiskResource(DiskResource diskResource);
     }
 
     HandlerRegistration addBeforeLoadHandler(BeforeLoadEvent.BeforeLoadHandler<FolderContentsLoadConfig> handler);
