+++ modules/ui-disk-resource-module/src/main/java/org/iplantc/de/diskResource/client/gin/DiskResourceGinModule.java
@@ -1,6 +1,7 @@
 package org.iplantc.de.diskResource.client.gin;
 
 import org.iplantc.de.client.models.diskResources.Folder;
+import org.iplantc.de.diskResource.DetailsView;
 import org.iplantc.de.diskResource.client.DiskResourceView;
 import org.iplantc.de.diskResource.client.GridView;
 import org.iplantc.de.diskResource.client.NavigationView;
@@ -20,6 +21,7 @@ import org.iplantc.de.diskResource.client.search.views.DiskResourceSearchField;
 import org.iplantc.de.diskResource.client.search.views.cells.DiskResourceQueryForm;
 import org.iplantc.de.diskResource.client.search.views.cells.DiskResourceSearchCell;
 import org.iplantc.de.diskResource.client.views.DiskResourceViewImpl;
+import org.iplantc.de.diskResource.client.views.details.DetailsViewImpl;
 import org.iplantc.de.diskResource.client.views.grid.GridViewImpl;
 import org.iplantc.de.diskResource.client.views.navigation.NavigationViewImpl;
 import org.iplantc.de.diskResource.client.views.toolbar.DiskResourceViewToolbarImpl;
@@ -95,6 +97,11 @@ public class DiskResourceGinModule extends AbstractGinModule {
         install(new GinFactoryModuleBuilder()
                     .implement(ToolbarView.Presenter.class, ToolbarViewPresenterImpl.class)
                     .build(ToolbarViewPresenterFactory.class));
+
+        // Details
+        install(new GinFactoryModuleBuilder()
+                    .implement(DetailsView.class, DetailsViewImpl.class)
+                    .build(DetailsViewFactory.class));
     }
