+++ modules/ui-disk-resource-module/src/main/java/org/iplantc/de/diskResource/client/gin/factory/DetailsViewFactory.java
@@ -1,0 +1,11 @@
+package org.iplantc.de.diskResource.client.gin.factory;
+
+import org.iplantc.de.diskResource.DetailsView;
+
+/**
+ * Created by jstroot on 2/2/15.
+ * @author jstroot
+ */
+public interface DetailsViewFactory {
+    DetailsView create(DetailsView.Presenter presenter);
+}
