+++ modules/ui-disk-resource-module/src/main/java/org/iplantc/de/diskResource/client/presenters/proxy/SelectDiskResourceByIdStoreAddHandler.java
@@ -45,7 +45,7 @@ public final class SelectDiskResourceByIdStoreAddHandler implements StoreAddHand
                     Scheduler.get().scheduleDeferred(new ScheduledCommand() {
                         @Override
                         public void execute() {
-                            presenter.setSelectedDiskResources(diskResourcesToSelect);
+                            presenter.setSelectedDiskResourcesById(diskResourcesToSelect);
                         }
                     });
                     presenter.unRegisterHandler(this);
