+++ modules/ui-disk-resource-module/src/main/java/org/iplantc/de/diskResource/client/views/details/DetailsViewImpl.ui.xml
@@ -1,0 +1,7 @@
+<ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder'
+             xmlns:g='urn:import:com.google.gwt.user.client.ui'>
+    <!-- Hard code with widgets? Or HTML mixin? -->
+    <div>
+
+    </div>
+</ui:UiBinder>
