+++ modules/ui-resources-module/src/main/resources/org/iplantc/de/resources/client/messages/IplantErrorStrings.properties
@@ -85,7 +85,6 @@ resourcesContainAncestors = Cannot move an ancestor folder into one of its desce
 restoreDefaultMsg = Unable to restore selected files and folders.
 retrieveAppListingFailed = Failed to retrieve app group contents.
 retrieveFolderInfoFailed = Failed to retrieve folder contents.
-retrieveStatFailed = Unable to retrieve details.
 retrieveSavedQueryTemplatesFailed = Failed to retrieve saved filters.
 saveParamFailed = Unable to save this parameters as a file. Please try again.
 searchError = Unable to complete your search. Please try again.
